refactor: extract _sandbox from sdk.py
Move sandbox user setup, path lockdown, privilege drop, and verifier
hardening into a new _sandbox.py module: setup_sandbox_user,
build_priv_drop_cmd, lockdown_paths, harden_before_verify, plus the
supporting cast (_validate_locked_path, _resolve_locked_paths, defaults).

VERIFIER_ENV and CLEANUP_CMD move from SDK ClassVars to module constants
in _sandbox.py — sdk.py now reaches them only transitively through
harden_before_verify, so no back-import is needed.

sdk.py: 1092 → 908 lines. _sandbox.py: 232 lines. Tests updated to import
from the new module; 474/474 pass.
